move AgentInfo and GetAgentInfo
